upgrade DS creation and DSCategories Editing interface to make it clearer for the user
