+++ app/src/main/res/layout/fragment_display_dataset.xml
@@ -6,14 +6,12 @@
     android:layout_height="match_parent"
     tools:context=".view.dataset_editing.DisplayDatasetFragment">
 
-    <EditText
+    <TextView
         android:id="@+id/display_dataset_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:textSize="30sp"
-        android:autofillHints="Dataset Name"
-        android:hint="@string/dataset_name"
-        android:inputType="text"
+        android:text="@string/dataset_name"
         android:maxLines="1"
         android:textAlignment="center"
         app:layout_constraintBottom_toTopOf="@id/display_ds_guideline1"
